pyproject.toml
[project]
name = "agenteval-backend"
version = "0.1.0"
description = "FastAPI backend for the AI Agent evaluation platform"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "boto3>=1.35.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "google-genai>=1.39.0",
  "httpx>=0.28.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["tests"]

[tool.uv]
package = false
